requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
pymongo==4.3.3
python-dotenv==1.0.0
email-validator==2.0.0
python-magic==0.4.27
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.1
